+++ core/src/main/java/org/conceptoriented/bistro/core/EvaluatorCalc.java
@@ -1,0 +1,9 @@
+package org.conceptoriented.bistro.core;
+
+/**
+ * This method knows how to compute one output value given several input values as an array.
+ */
+@FunctionalInterface
+public interface EvaluatorCalc {
+    public Object eval(Object[] params) throws BistroError;
+}