+++ core/src/main/java/org/conceptoriented/bistro/core/EvaluatorRoll.java
@@ -1,7 +1,7 @@
 package org.conceptoriented.bistro.core;
 
 /**
- * This methods knows how to update aggregate given a group member described by parameters and its distance from the group center.
+ * This methods knows how to update an aggregate given a fact (group member) described by its parameters and its distance from the group center.
  */
 @FunctionalInterface
 public interface EvaluatorRoll {