+++ examples/src/main/java/org/conceptoriented/bistro/examples/Example2.java
@@ -59,7 +59,7 @@ public class Example2 {
         productsAmount.setDefaultValue(0.0); // It will be used as an initial value
         productsAmount.accu(
                 itemsProduct,
-                p -> (double)p[0] + (double)p[1], // [Amount] + [out]
+                (a,p) -> (double)p[0] + (double)a, // [Amount] + [out]
                 items.getColumn("Amount")
         );